+++ R/extract_het_var.R
@@ -1,0 +1,48 @@
+#' extract_nlme_variances
+#'
+#' @description Because nlme won't do it for you.
+#' @param model An appropriate mixed model
+#' @param ... Other arguments appropriate to the method
+#'
+#' @details For models with heterogeneous variance, i.e. that contain something
+#'   like varIdent(form = ~1|Group), nlme returns a result in the summary
+#'   regarding the variances that many do not know what to do with, nor likely
+#'   would be what they would want to report. I rarely do these models with
+#'   nlme, and have only played around with the varIdent case.   This function '
+#'   will save me a little trouble in demos I do, but I have no plans to do any
+#'   more with it.
+#' @return A vector of the estimates on the variance scale.
+#' @importFrom stats coef
+#'
+#' @examples
+#' library(nlme)
+#' model <- lme(distance ~ age + Sex, data = Orthodont, random = ~ 1|Subject,
+#' weights=varIdent(form = ~1|Sex))
+#' summary(model)
+#' extract_het_var(model)
+#'
+#' @export
+extract_het_var <- function(
+  model,
+  ...
+) {
+  if (!inherits(model, c('lme')))
+    stop('This only works for model objects from nlme at present.')
+
+  UseMethod('extract_het_var')
+}
+
+#' @export
+extract_het_var.lme <- function(model, ...) {
+  init = coef(model$modelStruct$varStruct, unconstrained = F)
+
+  out = (c(1.0, init) * model$sigma) ^ 2
+
+  reflev = attributes(model$modelStruct$varStruct)$groupNames[1]
+
+  names(out)[1] = reflev
+
+  data.frame(as.list(out))
+}
+
+