+++ R/globals.R
@@ -1,16 +1,4 @@
 # so that check ignores
-utils::globalVariables(
-  c('variance',
-    'coefficient',
-    'coefficient_2',
-    'StdDev',
-    'Variance',
-    'est.',
-    'group',
-    'intervals',
-    'lower',
-    'upper',
-    'var_prop',
-    'effect',
-    'value'
-    ))
+utils::globalVariables(c('variance', 'coefficient', 'coefficient_2', 'StdDev',
+                         'Variance', 'est.', 'group', 'intervals', 'lower',
+                         'upper', 'var_prop', 'Estimate', 'sd'))