+++ NEWS.md
@@ -1,3 +1,11 @@
+# mixedup (development version)
+
+* Minor edit to fix ordering and names for extract_vc.lme
+
+# mixedup 0.2.1
+
+* Add method to extract_vc for brmsfit objects
+
 # mixedup 0.2.0
 
 * Core functions for primary packages working